+++ packages/omniston-sdk-react/src/index.ts
@@ -2,7 +2,9 @@ export * from "@ston-fi/omniston-sdk";
 
 export * from "./OmnistonProvider";
 export * from "./useBuildTransfer";
+export * from "./useBuildWithdrawal";
 export * from "./useConnectionStatus";
+export * from "./useEscrowList";
 export * from "./useOmniston";
 export * from "./useRfq";
 export * from "./useTrackTrade";